Bump azure SDK to 1.0/1.1, cache Entra credential per connection, split auth into entra.rs
