+++ mssql-odbc/Cargo.toml
@@ -10,11 +10,11 @@ crate-type = ["cdylib"]
 
 [dependencies]
 mssql-tds = { path = "../mssql-tds" }
-tokio = { version = "1", features = ["rt-multi-thread"] }
+tokio = { version = "1", features = ["rt-multi-thread", "sync"] }
 tracing = { version = "0.1.43", features = ["log"] }
 tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt"] }
 async-trait = "0.1"
-azure_identity = "0.25.0"
-azure_core = "0.25.0"
+azure_identity = "1.0"
+azure_core = "1.1"
 
 [build-dependencies]
